Move all but the public one into detail directory
