+++ include/podio/detail/AssociationCollectionIterator.h
@@ -1,5 +1,5 @@
-#ifndef PODIO_ASSOCIATIONCOLLECTIONITERATOR_H
-#define PODIO_ASSOCIATIONCOLLECTIONITERATOR_H
+#ifndef PODIO_DETAIL_ASSOCIATIONCOLLECTIONITERATOR_H
+#define PODIO_DETAIL_ASSOCIATIONCOLLECTIONITERATOR_H
 
 #include "podio/detail/AssociationCollectionData.h"
 #include "podio/detail/AssociationFwd.h"
@@ -43,4 +43,4 @@ private:
 };
 } // namespace podio
 
-#endif // PODIO_ASSOCIATIONCOLLECTIONITERATOR_H
+#endif // PODIO_DETAIL_ASSOCIATIONCOLLECTIONITERATOR_H
